+++ ds.py
@@ -1,4 +1,0 @@
-from downstream_farmer.shell import main
-
-if __name__ == '__main__':
-    main()